update frontend deps + fix some backend reqs not working
storytime:

so, for those who don't know, we are using vite's proxy server
functionality to reroute requests to /api to the backend server during
development. but, only for single-level-deep requests (i.e. /api/player,
but not /api/player/1), it would just... not forward the cookies (and
hence the session state)???? turns out ADDING A FUCKING SLASH AT THE END
OF THE URL FETCHED FIXES IT??? WTF??? this literally took 5 goddamn
hours to figure out, and i want to strangle someone