make a radical pivot in deployment mechanism
because i cannot for the life of me get nginx to work, we have pivoted
to a completely different way to deploy the application - make the
sveltekit app a static site, and serve it from within flask.

this reduces the complexity of the deployment, in the case we need to
deploy the website. still have no idea if any of the free flask
deployment sites allow usage of the filesystem (since we use sqlite)