nuke python
ok, ok, hear me out.

the issue with python was deployment. it's insane how difficult it is to
deploy this app with a dedicated frontend and backend. i tried nginx,
but it didn't work. then i tried serving the sveltekit app statically
from within flask, but it has dynamic routes, so it doesnt work. i have
decided that it's simply easier to rewrite the entirety of the python
api in sveltekit server routes. it only took like 3 hours, and now we
can actually deploy the damn thing... and for free. (thanks, vercel!)

also, while rewriting, i took care of some bugs that i found along the
way. like, absolutely terrible bugs. apparently, in either the team or
player endpoints, there was a function being used that calculated stat
rates for rows of a completely different table? crazy.

also, we're using a different db provider now. it's still sqlite for the
most part, but we'll be able to deploy it for free on turso.