do 404 pages properly
this is just silly... the reason the lockdown was registering a success
was because the /api/matches/lockdown endpoint was returning a 200, even
though it didn't exist... because of the [...404] matcher at the top
level of the project.

i have since found that i've been doing sveltekit 404 pages completely
wrong this entire time, and it's been corrected now. as a minor
consequence, the error page looks slightly different, but we can live
with that (hopefully)